Move IndexParameters to separate file
Also move some I/O functions to a separate file as they are used from both
StrobemerIndex and IndexParameters.
