+++ src/main.cpp
@@ -66,8 +66,7 @@ void log_parameters(const IndexParameters& index_parameters, const mapping_param
 }
 
 int run_strobealign(int argc, char **argv) {
-    CommandLineOptions opt;
-    opt = parse_command_line_arguments(argc, argv);
+    auto opt = parse_command_line_arguments(argc, argv);
 
     logger.set_level(opt.verbose ? LOG_DEBUG : LOG_INFO);
     logger.info() << std::setprecision(2) << std::fixed;
